requirements.txt
rockit-meco==0.1.28
casadi==3.5.5
matplotlib
